Reorganize Camera types
